feat: make a websites checker concurrent
Add concurrency to a website function that checks site availability.
To do that, it was used the concept of go routines, but go routines
writting in a map, is a bad thing, because map cant be written
at the same time by more then one process, so it was used chan.
Channels can receive and send information. With a struct, first
it was written, and then passed the result to the map.